Move PDBxFile into `cif.py`
